#1656 run hydrators/ direct updates
Replace UpdateOrderStateToSubmitted with SubmitOrder action

Consolidated the order submission logic by replacing UpdateOrderStateToSubmitted with SubmitOrder across the codebase. This simplifies the implementation while maintaining the same functionality, with added updates to affected routes, tests, and hydration logic.
